requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.5.2
astunparse==1.6.3
azure-core==1.32.0
azure-storage-blob==12.24.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
flatbuffers==24.3.25
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.68.1
h11==0.14.0
h5py==3.11.0
idna==3.10
importlib-metadata==8.5.0
isodate==0.7.2
keras==2.13.1
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.4.0
packaging==24.2
protobuf==4.25.5
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic-core==2.27.1
python-dotenv==1.0.1
python-multipart==0.0.19
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-cpu==2.9.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
werkzeug==3.0.6
wrapt==1.17.0
zipp==3.20.2
